fix: move pingHealthEndpoint functionality to health_check controller file
